refactor: seperate discount logics from cart
