+++ src/refactoring/hooks/useDiscount.ts
@@ -1,0 +1,17 @@
+import { CartItem, Discount } from "../../types";
+import { getAppliedDiscount, getMaxDiscount } from "../models/discount";
+
+export const useDiscount = () => {
+    const calculateMaxDiscount = (discounts: Discount[]) => {
+        return getMaxDiscount(discounts);
+    };
+    
+    const calculateDiscount = (item: CartItem) => {
+        return getAppliedDiscount(item);
+    };
+
+    return {
+        calculateMaxDiscount,
+        calculateDiscount,
+    }
+};
