+++ src/refactoring/models/cart.ts
@@ -1,4 +1,4 @@
-import { CartItem, Coupon, Discount, Product } from '../../types';
+import { CartItem, Coupon, Product } from '../../types';
 
 export const calculateItemTotal = (item: CartItem) => {
     const discountRate = getMaxApplicableDiscount(item);
@@ -76,23 +76,8 @@ export const updateCartItemQuantity = (cart: CartItem[], productId: string, newQ
     return newCart;
 };
 
-export const getMaxDiscount = (discounts: Discount[]) => {
-    return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
-};
-
 export const getRemainingStock = (cart: CartItem[], product: Product) => {
     const cartItem = cart.find((item) => item.product.id === product.id);
     return product.stock - (cartItem?.quantity || 0);
 };
 
-export const getAppliedDiscount = (item: CartItem) => {
-    const { discounts } = item.product;
-    const { quantity } = item;
-    let appliedDiscount = 0;
-    for (const discount of discounts) {
-        if (quantity >= discount.quantity) {
-            appliedDiscount = Math.max(appliedDiscount, discount.rate);
-        }
-    }
-    return appliedDiscount;
-};
